+++ docs/changelog.rst
@@ -15,6 +15,7 @@ The format is based on `Keep a Changelog`_ and this project adheres to
 **Added:**
 
 * Add client parameter checks and error handling
+* Read instance information from a configuration file
 
 
 `v0.1.0 <https://github.com/shaarli/python-shaarli-client/releases/tag/v0.1.0>`_ - 2017-03-12